Simplify setup.py
Current version is far too much boilerplate for a plugin
Its obsolete with the change to gulp, and has proven to be very unreliable.

We need to find something more simple that does not feel like black magic

Delete old sample_plugin. It is obsolete, now we have good examples
with waterfall and console

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>