buildbot_pkg: return valid version instead of latest
Yet we return a big number to make sure this is not confused with a real version

There are some use cases uncovered by the current method:

If a developer clones the source tree without fetching the tags, git describe
will not find the latest tag.

This would be incorrect to prevent installation in those cases.