Relax the check for the ssh executable name in integration tests.
It can be (and now actually is) "/usr/bin/ssh" instead of just "ssh", so use
endswith("ssh") instead of checking for equality.

Even this wouldn't pass under Windows where the executable is "ssh.exe", so it
would be arguably better to remove this check entirely, but until this test is
ran under Windows it can be kept.