serialize remoteCommand
Remotecommands updates are started in order, but are not garanteed to finish in order
We introduce a lock to make sure everything is serialized.

With buildbot nine, lots of the updates are doing asyncrhnous stuff like log update
and thus are definitly not finishing in order, thus a whole new class of weird behaviour.

Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>