Fix aliasing of codebase_ss
This variable was being set to {} in addBuild, but as a variable local
to addBuild, not to the test.  Switcihng that to codebase_ss.clear()
fixes the issue by not using assignment.  This patch also makes this
variable a dict of sourcestamps by codebase, rather than a dict of lists
of sourcestamps -- it's invalid to have multiple sourcestamps for the
same codebase.