Add methods for querying "attribute groups"
An "attribute group" is a group of attributes with a common name prefix.
The "mode" attribute group contains two members: full and incremental.
The full and incremental names must have the same "mode_" prefix, making
the method names "mode_full" and "mode_incremental".

The _hasAttrGroupMember and _getAttrGroupMember are the attribute group
equivalents of hasattr and getattr. The _listAttrGroupMembers exists so
options can be returned in error messages to help people know what the
valid options are without checking the docs or source.