status: URL-quote child help resource links
On the JSON API help page, URL-quote child names to prevent broken
links to child pages where the child nodes contains special characters.

Note that this cannot be done with Jinja2's urlencode filter, since that
doesn't escape slashes, which is a perfectly valid character in (for
example) builder names.

Signed-off-by: Anton Lofgren <alofgren@op5.com>