remoteInterruptCommand: add builderName argument
The rpc method can be directed to two type of objects:
- slave.buildslave.base.BotBase, or
- SlaveBuilder, that is contained on the previous object
  inside a dictionary {buildername: SlaveBuilder}.

Currently, the remote interruptCommand rpc is directed
to the slave.buildslave.base.Botbase. However this
object does not have any interruptCommand method,
only the SlaveBuilder has it.

Moreover, the current prototype of the method remoteInterruptCommand
- (commandId, why)
does not permit to find the associated SlaveBuidler to which
the rpc should be directed to.

This commit implements:
- a change of the prototype to the remoteInterruptCommand method
  to add the builderName, i.e (builderName, commandId, why)
- a redirection of the interruptCommand rpc, to the SlaveBuilder.