Introduce master log compression.
Log in the db is stored as chunck of text.
The implementation expected a compress operation to be done at the end
of the step. But it will be costly in time/db operations with db
fragmentation risk.

This patch suggest to do live compression when chunck size to insert
is smaller once compressed.
The compress operation will "just" have then to optimize/regroup
all contiguous text chunk, compress them and replace them in the db.

Running that algorithm 50x on the "pyflakes" job, shows that:
- we stored 0.583 MB of log in db instead of 3.079 MB
- space saving ration: 81.05%

Signed-off-by: Sebastien Fusilier <sebastien.fusilier@intel.com>