Introduce master log compression.
Log in the db is stored as chunck of text.
The implementation expected a compress operation to be done at the end
of the step. But it will be costly in time/db operations with db
fragmentation risk.

This patch suggest to do live compression when chunck size to insert
is smaller once compressed.
The compress operation will "just" have then to optimize/regroup
all contiguous text chunk, compress them and replace them in the db.
Compression algorithm is configurable in the master cfg.

Signed-off-by: Sebastien Fusilier <sebastien.fusilier@intel.com>