Fix permission errors with mock
mock builds fail with a permission error, even if the user running the build
slave was added to the mock group and packages can be built from
the console.

Using /usr/bin/mock as the following error message says fix the problem:

ERROR: [Errno 1] Operation not permitted
ERROR: The most common cause for this error is trying to run /usr/sbin/mock as an unprivileged user.
ERROR: Check your path to make sure that /usr/bin/ is listed before /usr/sbin, or manually run /usr/bin/mock to see if that fixes this problem.