service.py: make clustered service/scheduler manager an async service
The integration tests identify during some executions,
a race condition regarding AnyBranchSchedulers:
- a master starts, with startService
- the AnyBranchSheduler is started by means of
  the SchedulerManager (a service of the master)
  but returns before having started (as the scheduler
  is not an AsynService)
- one change, C, is added,
- the AnyBranchScheduler listens to new change events
  and is now started.

In that situation, the change C will not trigger any build
related to the AnyBranchScheduler configuration.

This commit makes clustered service and the related scheduler
manager become an asyncService, so that their startService
returns a deferred that is fired once they're activated only.

The reconfigService may still not be working though, which
may require to change the SchedulerManager into a
BuildbotServiceManager.