When requesting builds from Data API, properties can be filtered
Currently when requesting on Data API for builds (api/v2/builds),
each item (build) returned from the JSON structure,
does not contain the "properties" field.

The drawback of it, is that one's got to request for each build its properties,
which is costly in terms of performance and network usage
(in particular on big range of data).

It should be computed and served from the back-end (Data API).

With this patch, one can request builds properties like this:

- api/v2/builds/1?property=* (return the whole build's properties)
- api/v2/builds/1?property=slavename&property=user (return only desired properties)
- api/v2/builds?property=* (return the whole build's properties)
- api/v2/builds?property=slavename&property=user (return only desired properties)

Fields & Filters can be combined (limit, offset, order, ...)

Important: If field(s) is/are specified properties can be retrieved only if field
``properties`` is specified.

- api/v2/builds?field=buildid&field=properties&property=slavename&property=user

Note: By default, none properties are returned to avoid breaking API.

Unit tests & documentation updated accordingly into another commits.

Signed-off-by: Namgyal Brisson <namat4css@gmail.com>