Fix try scheduler, mark test as no longer flaky
This flaky test wasn't being run, and as a result we weren't seeing
the fact that the scheduler was subscribing to a nonexistent message
stream.

This also fixes a loop that would bail out on seeing just one
duplicate build, rather than simply skipping that build.  I doubt this
ever caused problems, but this is more correct.