Get results of build in StatusPush handler
The 'buildFinished' event indicates to the StatusPush callback that
the build is complete.  Considering that all steps of the build have
completed, it's useful to know whether they all completed
successfully or otherwise.

This is easy to accomplish because StatusPush::buildFinished already
has the success/error value available. Simply push it with the rest of
the event onto the queue.