Consistently refer to renderables as such in the docs
The previous version of the page can't seem to make up its mind about whether
to call them placeholders, annotations or renderables. Use the latter term
consistently, since it coincides with the interface name. The term
"placeholder" is now used only to refer to the %s-like constructs in format
strings.

There are also some collateral changes:

* explain more about renderables upfront, instead of delaying until the
  "Custom Renderables" section;

* clarify the text in some spots where it wasn't clear enough that
  arbitrary renderables are accepted.