master: source: git: Respect clobberOnFailure and retryFetch for git-fetch
The purpose of the clobberOnFailure and retryFetch parameters is either
retry the git-fetch command or perform a full clone of the repository
if the git-fetch failed. However, we only did that for git-reset and
not git-fetch so rework the code to pass this parameter to the git-fetch
step as well.