convert data api spec to raml
raml is a metalanguage to describe REST apis. http://raml.org/

The claim is to be the successor from swagger and blueprint, with focus being
on a easy to read source code.

My goal is is to evntually have all the rtype and most of data api doc
written in raml, with a bunch of scripts to convert the raml into

- sphinx doc for resource types and data api (regenerate the same doc as we currently have)
- sphinx doc for rest api (we don't have that right now, and ask our users to manually derive it from data api)
- sphinx doc for mq (which currently is mostly outdated, as it is derived from the rtype api http://trac.buildbot.net/ticket/3214)

All the raml file in this commit are purely derived from the typing metadata associated with endpoints declaration in the data api.

The goal is to enhance the raml files so that they include the rst doc we currently have in rtype-*.rst
and to have a unit test to automatically verify coherency b/w raml doc, and data/*.py

Another usage of this is also to generate metadata for the data-module, which currently hardcodes some stuff in the coffee.