create tables using create_all() + fix related issues
When tables are created using create_all() not only tables created,
but also SQLAlchemy runs its hacks, e.g. for cyclic referencew with
use_alter=True.

With this change some tables sets are not created in databases that
enforces relations between tables (e.g. PostgreSQL), because that set
of tables contained references on not created tables.