reexport .ai files using Inkscape
with "Replace PDF fonts by closes-named installed fonts" option
disabled.

Now SVG images are almost the same, but with text as text, not path.
(Though capital and lower letters are made as different words though.)

However text is using proprietary fonts "Myriad Pro" and "Myriad Pro
Cond" fonts which are not available on the most platforms.