Install [docs] dependencies in setup.py
Remove the pip install command from the docs Makefile used for installing the sphinx distribution.
Add the distribution to the extras in setup.py so that it can be installed if required by the user.
Add condition in the conf.py file to check if the [docs] dependencies are installed. If not, suitable error is displayed.

fixes #3435