compress logs in a transaction
During the log compression process, there is a time where all the log chunks are deleted
If while that a process reads the log, it will find an empty log.

This happens on integrations tests making them flaky.

This can theorically also happen for reporters even if we did not get any bug reports on that

What we do is to make the delete + re-insert a transaction, so that any read in between will get the old un-compressed version of the chunks (which is an okay answer)

I can reproduce the bug easily with a combination of sleep 2 after delete, and sleep 1 at the begining of getLogLines's thd,
but I dont want to cripple the code to add the necessary syncrhonisation hooks required to unit test this bug.