wrap commands long description with textwrap.dedent
Twisted reformats description text by removing newlines and rewrapping
text (at least Twisted 16.1.0).

Was:

```
$ buildbot-worker create-worker --help
Usage:    buildbot-worker create-worker [options] <basedir> <master> <name> <passwd>
...
This command creates a buildbot worker directory and buildbot.tac     file.
The bot will use the <name> and <passwd> arguments to authenticate     itself
when connecting to the master. All commands are run in a     build-specific
subdirectory of <basedir>. <master> is a string of the     form
'hostname[:port]', and specifies where the buildmaster can be reached.     port
defaults to 9989      The appropriate values for <name>, <passwd>, and <master>
should be     provided to you by the buildmaster administrator. You must choose
<basedir>     yourself.
```

Now:

```
This command creates a buildbot worker directory and buildbot.tac file. The bot
will use the <name> and <passwd> arguments to authenticate itself when
connecting to the master. All commands are run in a build-specific subdirectory
of <basedir>. <master> is a string of the form 'hostname[:port]', and specifies
where the buildmaster can be reached. port defaults to 9989.  The appropriate
values for <name>, <passwd>, and <master> should be provided to you by the
buildmaster administrator. You must choose <basedir> yourself.
```

This is better, but still not the best result.
This may should be fixed by <https://twistedmatrix.com/trac/ticket/1601>.