Update remote error handling to relay remote errors
Previously, when a remote's remote (e.g. master) failed during an operation
(such as uploading a file), the remote received a CopiedFailure. The remote
would not not properly reconstruct the failure to pass onto master. This led
the remote raising additional exceptions rather than properly terminating the
current command.

With this approach, the traceback is not passed (although a copy exists in the
original remote's logs, but the error is properly passed through the remote back
to master.