buildstep: finish unfinished logs
The previous version did not call finish on
unfinished (log.finished == False) logs. Such
situations happened among others with logs not attached to
a RemoteCommand, like for example stdio on
a MasterShellCommand.

Not calling finish had at least, two side effects:
- the 'complete' of related log lines attribute was not set to 1
  in the database,
- master.data.updates.compressLog was not called on the final
  state of the log.

For these reasons, this commit calls 'finish' on unfinised
logs before ending the step. The result of the step is set
to EXCEPTION if an error occured during that operation.

Change-Id: I30ebfc51ffac12cab22f4864494b3a334a202dac