.travis.yml: put lint tests first
More than 20 travis jobs are currently
executed to validate a pull request. Travis
however doesn't start the 20 builds in parallel.

The link tests are therefore executed after
previous ones are finished.
To avoid situations where all the unit tests are ok,
but one needs to wait 20 minutes to be aware
of a lint issue (incorrect order of import), which is quite likely
to happen, this commit puts the lint test first.

Change-Id: Id57d6d9fec4742cc7ea768d888e058835ebb2113