Rather than trying to keep track of building WFBs, just ask them.
The changes to have substantiation run as part of the build means that
when a substantiation failed, the list of things building was cleared
(by insubstantiate), and then the triggering build finishing would try
to remove an entry, which would then fail.

Rather than try to keep this in sync, just ask.