Make EC2 spot instance tests use spot instances
Repair the spot instance tests to actually attempt to use
spot instances.  EC2LatentWorker()'s internal method _start_instance()
starts a normal EC2 On-Demand instance, not a Spot Instance.  It
is the _request_spot_instance() method that requests a spot instance,
so change the tests to use that.

Also, region and placement must be set for EC2LatentWorker when
spot instances are used, because the AvailabilityZone zone cannot
be None in later AWS calls.

This is expected to fail in testing on the step that requests the
spot instance historical data.