Remove the ec2 spot price retry on price-too-low
The code to retry spot requests with a bid price increased
a multiple was well intentioned, but based on a faulty understanding
of AWS spot pricing.  AWS only charges the _current_ spot price
when a spot instance is running.  The prices varies over time, and
if the price exceeds the bid price, the instance is stopped.

There is little or no downside to bidding the highest price that
one is willing to pay right at the outset, because the price charged
is the current price, not the bid price.

This removes the EC2LatentWorker "retry" and "retry_price_adjustment"
options.