be more flexible in getResponsibleUsers
Normally properties['owner'] shall be a string.
We got some users reporting it is a list:
https://github.com/buildbot/buildbot/pull/2284
I couldn't find in the code any evidence where owner would be set to a list.

We don't have much more information, so lets log a message in that case in the hope we get more
reports