Have proper Image, Volume, and Snapshot types
The next feature to automatically determine the volume size needs to be
able to do basic checking of the image, volume, or snapshot object. This
requires the image to be more than a "UUID" string.

Turn the Images class into an ItemManager that can be used to manage
images, volumes, or snapshots. The ItemManager class supports only the
bare minimum of API that novaclient has.

Create a generic Item class and subclass that to make the Image, Volume,
and Snapshot types. Add a default image, volume, and snapshot to not
force every test to do the same when needed.