Fix the cancel and stop story
Cancel of a buildrequest was not working in the buildrequest page.

When there are tons of buildrequests in the queue this is very painful to stop them all manually

So we also add a feature that was in buildbot eight cancel whole queue. Unlike in eight the whole queue means buildrequests *and* running builds for the sake of simplicity

We can still change the code to add a separate button to avoid killing running builds.