don't disconnect if building when doing soft disconnect
At master reconfiguration, some worker might be disconfigured.
This is not a good reason to kill their builds.

This has the effect to slow down the reconfig until all disconfigured worker have finished their work.
The same goes for master shutdown

master will now kill the running builds at stop rather than just killing the workers
(in subsequent commits)

latent workerforbuilder: put the worker in state DETACHED when it is substanciating

In the case where the worker is substanciating, it is safe to kill the build at that point
as no work as been done yet