Fix BitbucketPullrequestPoller change detection
Currently the BitbucketPullrequestPoller treats all pull requests as
if they have been updated, rather than filtering them properly. This is
because the existing check compares the short hash of the incoming
changeset with the full hash of the current revision.

This patch changes the behaviour to compare the short hashes of the
incoming and current revisions.

The Bitbucket API returns short hashes for a number of queries. The
test data has been updated to reflect this.

fixes #3597

Signed-off-by: Sam Bristow <sam.bristow@gmail.com>

Explicitly use short hashes when comparing PR revisions

This makes the version comparision more robust in case Atlassian changes the
Bitbucket PR API response to return full hashes in future.

Signed-off-by: Sam Bristow <sam.bristow@gmail.com>