Remove use of class level variables in ListResult conflicting with __slots__.
In Python 3, this was triggering this error:

builtins.ValueError: 'offset' in __slots__ conflicts with class variable

In Python 2, if you defined slots and after that defined a property
with the same name the slots descriptor was replaced silently. In
Python 3 this isn't allowed anymore.

Using __slots__ may offer memory advantages for certain types
of objects.

See:
https://github.com/paramiko/paramiko/issues/103#issuecomment-15127799
http://bugs.python.org/issue12766
http://book.pythontips.com/en/latest/__slots__magic.html