buildbotNetUsageData implementation
This feature sends data to the buildbot.net statistics server
This will allow the buildbot developers to better understand the buildbot
installation base and thus to better prioritize new developements

This is an opt-out design.

By default it sends basic information, with a warning in the log.

User can the opt-out for no information sent or opt-in for more information.