Call existing disconnect method if there should be no wait
When build_wait_timeout is greater than 0, a timer is set to call
_soft_disconnect. That method has some checking and then does a proper
disconnect and insubstantiate. Furthermore, within the insubstantiate
method, it already checks if there are builds pending to cause the
worker to substantiate again immediately. When build_wait_timeout is 0,
the timer doesn't need to be set so call what it would otherwise call.
This will ensure the two cases stay in sync.

Fixes trac #3610.