force latent worker to restart if build_wait_timeout is 0
build_wait_timeout = 0 kind of worked in real reactor, but sometimes some build start, and then are stopped right after

This is because of a deferred race condition which is not happening on the fake reactor

baseWorker.buildFinished will start the process of searching new build
but latentWorker.buildFinished will insubstantiate in parrallel
depending on how fast the db will return the unclaimed buildrequest, a new build can start before the insubstantiation.

by calling baseWorker.buildFinished after doing the insubstantiate, we make sure that this worker will not be selected

This commit also contains test for this case, and fix for LocalSlave to correctly handle its lifecycle.