Replace cStringIO.StringIO with io.BytesIO and io.StringIO for Python 3
On Python 2, sys.stdin.read() returns bytes, so we must use io.BytesIO there
On Python 3, sys.std.read() returns str, so we must use io.StringIO there.

In addition md5.update() requires an argument of type bytes, so if
the job variable is a unicode string (especially on Python 3), we must encode it.