resultSpec: apply resultSpec in the db layer when it is simple
There is often a quite simple mapping from data to db api

This commit teaches resultSpec to automatically convert itself into sqlalchemy filters, order and paging api.

When this is not possible, the resultSpec will still be processed in the data layer as a postprocessing.

This avoids the UI to start being very slow when there is a big number of builds in the db.
Now the paging is directly handled by the db layer, the requests can be very quick even if the db is crowded