Update StashStatusPush to allow customizing payload
The JSON object has three free-form fields: key, name, and description.
Previously, only the key was sent and it was hard-coded to be the name
of the builder. Add parameters to allow setting the key, name, and
description. Like the GitHub reporter, support a description for when
the build has started/in-progress or finished. These parameters support
being rendered.