buildPushesWithNoCommits also rebuilds known branch after update
The buildPushesWithNoCommits in gitpoller.py determines if a push
on a new branch or update of an already known branch with already
known commits should trigger a build.

Until this, it worked fine for new branches, but would not trigger
a build if, for instance, you fast-forwarded your master branch onto
a topic branch for which a build had already happened. This commit
fixes this behaviour.