fix logviewer bad scrolling
While the scrollviewport is updating, there is a chance
that the overall size of elements decrease.

If we are scrolled to the bottom, then the scroll viewport will automatically
go up to fit new smaller size, and not go back down when the new elements are inserted.

Thus we will not see the end of the log anymore.
We add a parent div, which is always of the maximum log size