BuildbotServiceManager: support reconfiguration for legacy services
We have a lot of legacy services (Schedulers and ChangeSources)
which do not support reconfigService, and are not reconfigurable.

Porting everything to reconfigurable is a lot of work, mainly in the unit tests
as they must properly construct a service hierarchy.

Rewritting everything makes few sense, so we add a method so that the
BuildbotServiceManager detects when the service do not support reconfiguration
and then just swap the old service with the new service, instead of reconfigure.
This is racy and not optimal, but much better than not reconfiguring (without warning).

Signed-off-by: Pierre Tardy <tardyp@gmail.com>