Added a try/except around a call to multiprocessing.cpu_count()
In some situations, the multiprocessing.cpu_count() function can raise
NotImplementedError. I.e. this can happen on Windows if the environment
will lack the 'NUMBER_OF_PROCESSORS' variable, or this variable will
contain a non-numeric character.

If this exception will occur, the worker will use a substitution value
of 1 CPU.