Sort the array of dictionaries representing changes before comparing them.
In P4Souce._poll(), the dictionary of changes is populated by:

    branch_files = {}
    ...

    for branch in branch_files:
            addChange()

Since the order of items in a dictionary is not necessarily the
same in Python 2 and Python 3,
the order of branches in the branch_files dictionary is not
the same between Python 2 and Python 3.
This means that the order of changes in the array
is not the same between Python 2 and 3.

For the purposes of this test, we sort the array of changes
by branch when comparing them.