Use JWT for storing session data
Annoying problem with buildbot is that the session are lost when
the master is reboot.
With twisted web sessions are stored in memory, and the API being fully synchronous,
there is no sane way to store it in the db.

JWT is a technic which stores the session state within a cookie,
and sign it with secret key backed hmac.

We store the session secret in the db, and this is the only thing that we
need to store in the db.

Previously when you logged into one master, your cookie was invalid for the other master, now, it is valid for every master

With this implementation we can safely implement multimaster web UI
without the need for IPHash load balancing.