fix vacuum for python 3.6
Changed in version 3.6: sqlite3 used to implicitly commit an open transaction before DDL statements. This is no longer the case.

I removed all the actions of the cleanupdb command, and sqlite3 still think I am in a transaction.
looking at Connection.in_transaction, it returns False.

So I am not sure if it is a bug or something outside of my understanding.