Allow selecting fields and ordering descending at the same time
REST API allows to specify fields to return (via "field=" query param(s)) and
sorting order (via "order=" query param(s)); it also requires that order
fields are present in fields list. When order is descending field name in
"order=" param is to be prepended with "-", and this expression is an invalid
field name which cannot be found in fields list (note that prepending "-" to
field name in "field=" param is prohibited).

Strip leading "-" from order fields when checking whether all order fields
are also present in fields list. This would allow specifying fields to return
and sort in descending order at the same time.