Force all the print()'s  in printBuild() and printLog() to be unicode, and use io.StringIO() on Python 2 and Python 3.
Some of the parameters being passed to the print()'s
such build['state_string'] were unicode even on Python 2.
Mixing NativeStringIO and io.BytesIO for these print()'s on Python 2,
while using NativeStringIO and io.StringIO for these print()'s on Python 3
was not working out and causing test failures on Python 3.

This change fixes the tests on Python 2 and 3.