master: replace most str occurences by string_types
When configuring some services in master.cfg with:

    from __future__ import unicode_literals
    from buildbot.plugins.db import steps
    b = steps.ShellCommand(
        name='do stuff',
        command=['stuff'],
    )

there are errors with python 2:

    BuildStep name must be a string: u'do stuff'

BuildStep is only an example, there are other constructors that do not
accept unicode strings.

Since almost all internal API work on unicode strings, replace most
checks that a variable is a string by testing if it is an instance of
string_types (from future.utils).

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>